news_jungle_export/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.46.0",
    "nltk>=3.9.1",
    "openai>=1.63.2",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "streamlit>=1.42.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
]
